Increment patch version for new content.
Before this commit, consecutive commits on top of tag x.y.z would result
in a version of x.y.z-c, where c was the number of commits since the
tag.  This was not semver compatible, because a hyphen is used to denote
a prerelease, as discussed at:

https://semver.org/#spec-item-9

Particularly: "Pre-release versions have a lower precedence than the
associated normal version."

After this patch, we increment the patch version before designating it
as a pre-release. This leads to semver compatible versioning like:

5.5.37 < 5.5.38-1 < 5.5.38