Add +sha build metadata to the end of non-cannonical versions.
For all builds that are not directly on a tagged version, add 8 digits
of the commit SHA as build metadata described here:

https://semver.org/#spec-item-10

This allows us to distinguish between dev builds with the same number of
commits since a common tag, but different hashes.  This will be useful
if/when we persist build artifacts from try builds, and multiple
developers may generate the same semver build id, with different
content.