Build docs to gravity version build dir.
Instead of always building docs to ../build/docs, they are now built to
../build/$(VERSION)/docs when run from the top level makefile.  I'm a
bit worried this will confuse some tool that expects only binaries in
that directory.  Nonetheless, this allows for easier iteration on
different versions of the documentation.

As part of this, I needed to remove the $(BUILDDIR) target, because it
was constantly being updated by the build (rightfully so), and thus
re-triggering the $(DOCBOX_ID) target.